Add files via upload
The project features an automatic gate that identifies and closes for large vehicles, prompting them to divert. Upon detecting a heavy or tall vehicle approaching the gate, the system initiates a sequence of actions. Firstly, it promptly closes the gate, creating a barrier for the identified vehicle. Simultaneously, the system displays a message on an LCD screen, instructing the driver to take an alternative route ahead  ensuring the smooth flow of regular-sized vehicles, such as cars. 
              This project includes key components like an Arduino Uno, a servo motor responsible for gate movement, an LCD display for communication, and an ultrasonic sensor for precise measurement of vehicle height. The ultrasonic sensor plays a crucial role in the decision-making process, assessing whether a vehicle meets the specified height criteria.

The real-world applications of this system extend to areas with specific height restrictions such as underpasses, express highways, railway crossings,and tunnels